+++ docs/source/changes.md
@@ -12,6 +12,9 @@ releases are available on [PyPI](https://pypi.org/project/pytask) and
   (fixes #514). It also warns if the path is configured as a string and not a list of
   strings.
 - {pull}`519` raises an error when builtin functions are wrapped with
+  {func}`~pytask.task`. Closes {issue}`512`.pull
+- {pull}`521` raises an error message when imported functions are wrapped with
+  {func}`@task <pytask.task>` in a task module. Fixes {issue}`513`.
   {func}`~pytask.task`. Closes {issue}`512`.
 - {pull}`522` improves the issue templates.
 - {pull}`523` refactors `_pytask.console._get_file`.